feature(PBR): Normal mapping
Implemented normal mapping by reading a normal texture. Bitangent and tangent are computed after reading the vertices and a transform matrix is sent to the fragment shader. LearnOpenGL says that it would be better to do all the computations on the vertex shader, i'll probably give a better look at it when i'll have more than a single light in my scene.