feature(Pipeline): Refactored the pass system
A pass is a virtual class that gives common functions useful for all rendering. Then 2 interfaces are used to represent actual passes:  renderPasses that draw objects, and PostProcess pass that take a texture as input and edit it. I could use compute shader for post process instead of the rendering pipeline, and i should benchmark it too.