feature(Scene): Scene graph
Entities have children. This introduces local transform, though there is no way to set an object as child of another for the moment. Also refactored the UI to have a pseudo inspector for our entities and avoid having multiple guizmos active.

Since we do not have an array of entities in our editor anymore, i had to refactor the rendering code. I chose to non-recursively iterate on all the children of our entities to draw. I also unified the namings in our different passes, just in case